An attempt to appease MacOS's linker for versioning.
I picked 10.6 because that seems to be the earliest version my El Capitan machine supports. I have no idea if that's a good choice or not.

These version numbers seem to match the "About this mac" numbering (10.11.6 for my machine), not the kernel numbering (15.6.0 for the same machine). I have no idea what the deal is there.